backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
httpx>=0.27,<1
playwright>=1.49,<2
beautifulsoup4>=4.12,<5
lxml>=5.3,<6
shapely>=2.0,<3
pydantic>=2.10,<3
python-dotenv>=1.0,<2
